Hackathon02:27:26/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
duckdb==1.1.3
sqlparse==0.5.3
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
